refactor(player): decoupling and re-writing player code
This commit splits the responsibility of player across six new
scripts and improves the reliability of the code.

Additionally, it's been decided to remove the previous floaty
ghoul counter in favor of a standard score count to the top left
of the screen.

Also some minor changes elsewhere as and when I noticed things
that could be ripped out and/or improved.